shop items added

+++ admin/shop_image.php
@@ -12,8 +12,14 @@ include_once '../includes/admin-sidebar.php';
     <div class="block-header">
         <div class="row">
             <div class="col-lg-7 col-md-6 col-sm-12">
-                <h2 style="font-size: 16px">  <i class="fa fa-image mr-2"></i>Shop Item Picture               
+                      <?php
+                         if(null != filter_input(INPUT_GET, "n")){
+                               $n = filter_input(INPUT_GET, "n"); 
+                        $shops = $shop->get_menu_by_id($n); 
+                             ?>
+                <h2 style="font-size: 16px">  <i class="fa fa-image mr-2"></i><?php  echo $shops['name']. " "  ?> Picture               
                 </h2>
+                         <?php } ?>
             </div>
             <div class="col-lg-5 col-md-6 col-sm-12">
                 <button class="btn btn-white btn-icon btn-round float-right m-l-10" type="button" data-toggle="modal" data-target="#smallModal">
@@ -27,7 +33,12 @@ include_once '../includes/admin-sidebar.php';
             <div class="col-lg-12 col-md-12 col-sm-12">
                 <div class="card shadow ">
                     <div class="header">
-                        <h2><strong>List of </strong> Shop Item Pictures </h2>
+                        <?php
+                         if(null != filter_input(INPUT_GET, "n")){
+                          $n = filter_input(INPUT_GET, "n"); 
+                        $shops = $shop->get_menu_by_id($n); ?>
+                        <h2><strong><?php echo $shops['name']. ":"?> </strong><span style="font-size: 13px"> <?php echo $shops['description'] ?> </span></h2>
+                         <?php } ?>
                     </div>
                 </div>
             </div>
@@ -43,39 +54,32 @@ include_once '../includes/admin-sidebar.php';
                         </thead>                            
                         <tbody>
                             <?php
-//                            $count = 1;
-//                            $shops = $shop->get_all();
-//                            if ($shops != 0) {
-//                                foreach ($shops as $shop) {
+                            if (isset($_GET['n'])){
+                            $n = filter_input(INPUT_GET, "n");
+                            $count = 1;
+                            $shop_pics = $shop_pic->get_shop_pics_by_id($n);
+                            if ($shop_pics != 0) {
+                                foreach ($shop_pics as $shop_picture) {
                                     ?>
                                     <tr>
-                                        <td><?php // echo $count ?></td>
-                                        <td><?php // echo $menus['name'] ?></td>
+                                        <td><?php  echo $count ?></td>
+                                        <td><img src="../uploads/shop_item_img/<?php echo $shop_picture['picture'] ?>" width="110px"></td>
                                         <td>
-                                            <div class="dropdown" style="display: inline !important">
-                                                <a class="dropdown-toggle" href="#" type="button" data-toggle="dropdown" aria-expanded="false">
-                                                    <i class="fa fa-folder-open m-r-15 text-info" title="attachment"></i>
-                                                </a>
-                                                <div class="dropdown-menu">
-                                                    <a class="dropdown-item" href="shop_image?n=<?php // echo $shop['id'] ?>"><i class="fa fa-plus m-r-15 text-info"></i>Add new image</a>
-                                                    <a class="dropdown-item" href="#"><i class="fa fa-location-arrow m-r-15 text-info"></i>View all images</a>
-                                                </div>
-                                            </div>
-                                             <a href="update_shop?n=<?php // echo $menu['id']  ?>"><i class="fa fa-edit m-r-15 text-success" title="update shop item"></i></a>
                                             <div class="dropdown" style="display: inline !important">
                                                 <a class="dropdown-toggle" href="#" type="button" data-toggle="dropdown" aria-expanded="false">
                                                     <i class="fa fa-trash-alt m-r-15 text-danger" title="delete item"></i>
                                                 </a>
                                                 <div class="dropdown-menu">
-                                                    <a href="../controller/delete_menu?n=<?php // echo $menu['id']  ?> " class="dropdown-item"><i class="fa fa-trash m-r-15 text-danger"></i>Delete item</a>
+                                                    <a href="../controller/delete_menu?n=<?php // echo $menu['id']  ?> " class="dropdown-item"><i class="fa fa-trash m-r-15 text-danger"></i>Delete picture</a>
                                                 </div>
                                             </div>
                                         </td>
                                     </tr>
                                     <?php
-//                                    $count++;
-//
-//                            }
+                                    $count++;
+                                }
+                            }
+                            }
                             ?>
                         </tbody>
                     </table>
@@ -87,10 +91,7 @@ include_once '../includes/admin-sidebar.php';
 <div class="modal fade" id="smallModal" tabindex="-1" role="dialog">
     <div class="modal-dialog modal-md" role="document">
         <div class="modal-content">
-            <?php
-            if(null != filter_input(INPUT_GET, "n")){
-                $n = filter_input(INPUT_GET, "n"); ?>
-            <form action="../controller/add_shop_item_pics?n=<?php echo $n ?>" method="post">
+            <form action="../controller/add_shop_item_pics?n=<?php echo $n ?>" method="post" enctype="multipart/form-data">
                 <div class="modal-header">
                     <h4 class="title" id="smallModalLabel">Add New Shop Item Picture</h4>
                 </div>
@@ -105,15 +106,14 @@ include_once '../includes/admin-sidebar.php';
                             if (isset($_POST['name'])) {
 //                                echo $_POST['name'];
                             }
-                            ?>" name="name" class="form-control" placeholder="Select Picture">
+                            ?>" name="shop_item_photo" class="form-control" placeholder="Select Picture">
                         </div>
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <button type="submit" name="shop_def" class="btn btn-default btn-round waves-effect" style="background-color: #ed5a0b">ADD SHOP ITEM PICTURE</button>
+                    <button type="submit" name="shop_image" class="btn btn-default btn-round waves-effect" style="background-color: #ed5a0b">ADD SHOP ITEM PICTURE</button>
                 </div>
             </form>
-            <?php } ?>
         </div>
     </div>
 </div>
